CR123a to AA battery adapter for flashlight
An insert to allow a flashlight to use a CR123a battery in place of a
single AA battery. This is sized for an SK-68 LED flashlight that has a
slightly larger bore than normal, so that the CR123a fits (diameter is
larger than an AA battery). Add a screw or some wire as a conductor and
it's good to go.